# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JABEEN Investor Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Self-register as Investor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
        "403":
          description: Account deactivated
        "429":
          description: Too many attempts
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Renew session
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Renewal failed
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get own profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # --- DASHBOARD ---
...
  /projects/{projectId}/documents:
    get:
      operationId: listProjectDocuments
      tags: [documents]
      summary: List project documents
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document to a project
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUpload"
      responses:
        "201":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
  /documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document (admin only)
      parameters:
        - $ref: "#/components/parameters/documentId"
      responses:
        "204":
          description: Deleted
  /documents/{documentId}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Download document content
      parameters:
        - $ref: "#/components/parameters/documentId"
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # --- MESSAGES ---
  /projects/{projectId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List project messages
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Post a project message
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # --- INTERNAL NOTES ---
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
    documentId:
      name: documentId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    RegisterInput:
      type: object
      required: [fullName, email, password, companyName]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 160
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        companyName:
          type: string
          minLength: 2
          maxLength: 200
        title:
          type: string
          maxLength: 160
        phone:
          type: string
          maxLength: 40
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 160
        title:
          type: ["string", "null"]
          maxLength: 160
        companyName:
          type: string
          minLength: 2
          maxLength: 200
        phone:
          type: ["string", "null"]
          maxLength: 40

    # --- USER ---
    User:
      type: object
      required: [id, email, fullName, companyName, role, active, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        companyName:
          type: string
        title:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [investor, top-management, project-manager, administrator]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Document:
      type: object
      required: [id, projectId, fileName, contentType, size, uploaderId, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        updateId:
          type: ["integer", "null"]
        fileName:
          type: string
        contentType:
          type: string
        size:
          type: integer
        uploaderId:
          type: integer
        uploader:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    DocumentUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        updateId:
          type: integer

    # --- MESSAGE ---
    Message:
      type: object
      required: [id, projectId, authorId, authorRole, body, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        authorRole:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 4000

    # --- INTERNAL NOTE ---
...